Revert "refactor(Select): create BaseSelect components"
This reverts commit cf883d431a5d5239bd779dd97a424257cf802cb4.
