--- packages/react-ui-kit/src/Form/Select/BaseSelect/BaseSelectMenuList.styles.ts
@@ -1,40 +1,0 @@
-/*
- * Wire
- * Copyright (C) 2025 Wire Swiss GmbH
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see http://www.gnu.org/licenses/.
- *
- */
-
-import {CSSObject} from '@emotion/react';
-
-import {Theme} from '../../../Layout';
-
-export const headingContainerStyles = (theme: Theme): CSSObject => ({
-  display: 'flex',
-  justifyContent: 'space-between',
-  alignItems: 'center',
-  fontSize: theme.fontSizes.medium,
-  fontWeight: 600,
-  padding: '8px 16px',
-});
-
-export const closeButtonStyles: CSSObject = {
-  background: 'transparent',
-  border: 'none',
-  cursor: 'pointer',
-  padding: 0,
-  display: 'flex',
-  alignItems: 'center',
-};
