+++ packages/react-ui-kit/src/Form/index.ts
@@ -30,7 +30,7 @@ export * from './InputBlock';
 export * from './InputSubmitCombo';
 export * from './RangeInput';
 export * from './RoundIconButton';
-export * from './Select/Select';
+export * from './Select';
 export * from './ShakeBox';
 export * from './Switch';
 export * from './TextArea';
